moved old_prediction_controller to junk folder
